Moved base_dird_integration_test module to helpers.utils
