+++ integration_tests/suite/test_csv_backend.py
@@ -10,7 +10,7 @@ from hamcrest import assert_that, contains, contains_inanyorder, has_entries
 from .helpers.base import BaseDirdIntegrationTest, CSVWithMultipleDisplayTestCase
 from .helpers.config import new_csv_with_pipes_config
 from .helpers.constants import VALID_UUID
-from .base_dird_integration_test import absolute_file_name, BackendWrapper
+from .helpers.utils import absolute_file_name, BackendWrapper
 
 
 class _BaseCSVFileTestCase(unittest.TestCase):
