+++ integration_tests/suite/test_office365_backend.py
@@ -19,7 +19,7 @@ from unittest.mock import Mock
 from wazo_dird_client import Client as DirdClient
 from wazo_test_helpers.auth import AuthClient as AuthMock
 from wazo_test_helpers.hamcrest.raises import raises
-from .base_dird_integration_test import BackendWrapper
+from .helpers.utils import BackendWrapper
 from .helpers.base import DirdAssetRunningTestCase
 from .helpers.constants import (
     UNKNOWN_TENANT,
